mark-yank-urls: fix bug allowing shell to interpret the url
The problem was that the script was passing the url via exec_async
in a way that allowed the shell to interpret the url string.  If that
string contained characters like &, it would allow the URL to
accidentally, or maliciously, execute arbitrary commands.

This has been reported by a few people (thanks to Hans Dieter Pearcey,
Daniel Danner and Olof Johansson for reporting it) and I finally (got
off my butt and) looked at it.

Signed-off-by: Bart Trojanowski <bart@jukie.net>