Mark-yank-urls and Ctrl detection
I'm having some minor problems with the latest version; the ^p and ^n
don't seem to work. What I did with my not-so-good Perl skills I reverted
the control key detection to a previous version of it (patch attached).
This fixed the control key behaviour on my machine and, to my knowledge,
should not be affecting the behaviour on other computers either.

Signed-off-by: Bart Trojanowski <bart@jukie.net>