url_matcher for non-latin URLs (not default)
Contributed by worn_mind, with this message...

I'm using mark-yank-urls and i have a suggestion.
Today there are a lot of sites int the internet that uses national
alphabet in url, so i think it would be nice if mark-yank-urls could
handle it. So i modified your script(patch attached). Please think about
it, maybe it's worth to be added.

... it's not default since it's not tested.