Fix function call syntax error when using mouse clicks
  foo() vs foo{}

This fixes warnings like (as seen on urxvt's stdout/err), if one ever
gets the idea of middle clicking a link:

  Odd number of elements in anonymous hash at $HOME/.urxvt/mark-yank-urls line 104.
  Use of uninitialized value $scheme in hash element at $HOME/.urxvt/mark-yank-urls line 40.

This bug was introduced in my commit adding "ability to define handlers
per protocol". Because of the age of the branch, I will refrain from
rebasing and let this be a tombstone of one of my many mistakes
throughout the years. The expected behavior (processing links on mouse
button2 key release) was confirmed to be working with this change
applied.